adjusting log statements
